chore: 코드리뷰 적용
