--- app/src/main/java/com/flint/core/common/datastore/PreferencesManager.kt
@@ -1,43 +1,0 @@
-package com.flint.core.common.datastore
-
-import androidx.datastore.core.DataStore
-import androidx.datastore.preferences.core.Preferences
-import androidx.datastore.preferences.core.edit
-import androidx.datastore.preferences.core.stringPreferencesKey
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.map
-import javax.inject.Inject
-import javax.inject.Singleton
-
-@Singleton
-class PreferencesManager
-    @Inject
-    constructor(
-        private val dataStore: DataStore<Preferences>,
-    ) {
-        suspend fun saveString(
-            key: String,
-            value: String,
-        ) {
-            dataStore.edit { preferences ->
-                preferences[stringPreferencesKey(key)] = value
-            }
-        }
-
-        fun getString(key: String): Flow<String> =
-            dataStore.data.map { preferences ->
-                preferences[stringPreferencesKey(key)].orEmpty()
-            }
-
-        suspend fun removeString(key: String) {
-            dataStore.edit { preferences ->
-                preferences.remove(stringPreferencesKey(key))
-            }
-        }
-
-        suspend fun clearAll() {
-            dataStore.edit { preferences ->
-                preferences.clear()
-            }
-        }
-    }
