+++ app/src/main/java/com/flint/data/api/CollectionApi.kt
@@ -12,6 +12,6 @@ interface CollectionApi {
     // 컬렉션 상세 조회
 
     // 최근 본 컬렉션 목록 조회
-    @GET("api/v1/collections/recent")
+    @GET("/api/v1/collections/recent")
     suspend fun getRecentCollectionList(): BaseResponse<RecentCollectionListResponseDto>
 }
