+++ app/src/main/java/com/flint/data/api/ContentApi.kt
@@ -6,7 +6,7 @@ import retrofit2.http.GET
 
 interface ContentApi {
     // 북마크한 콘텐츠 목록 조회
-    @GET("api/v1/contents/bookmarks")
+    @GET("/api/v1/contents/bookmarks")
     suspend fun getBookmarkedContentList(): BaseResponse<BookmarkedContentListResponseDto>
 
     // 콘텐츠별 OTT 목록 조회
