--- app/src/main/java/com/flint/data/local/TokenStore.kt
@@ -1,3 +1,0 @@
-package com.flint.data.local
-
-class TokenStore
